+++ src/connected/MapPointInfoBox.tsx
@@ -22,48 +22,31 @@
  * SOFTWARE.
  */
 
-import React, { useMemo } from "react";
-import Box from "@mui/material/Box";
+import { connect } from "react-redux";
 
-import { getLabelsForRange } from "@/util/label";
-import { makeStyles } from "@/util/styles";
+import { AppState } from "@/states/appState";
+import _MapPointInfoBox from "@/components/MapPointInfoBox";
+import {
+  selectedDatasetSelector,
+  selectedDatasetTimeLabelSelector,
+  selectedServerSelector,
+  selectedVariableSelector,
+} from "@/selectors/controlSelectors";
 
-const styles = makeStyles({
-  container: {
-    fontSize: "x-small",
-    fontWeight: "bold",
-    width: "100%",
-    display: "flex",
-    flexWrap: "nowrap",
-    justifyContent: "space-between",
-    cursor: "pointer",
-  },
-});
+const mapStateToProps = (state: AppState) => {
+  return {
+    enabled: state.controlState.mapPointInfoBoxEnabled,
+    serverUrl: selectedServerSelector(state).url,
+    dataset: selectedDatasetSelector(state),
+    variable: selectedVariableSelector(state),
+    time: selectedDatasetTimeLabelSelector(state),
+  };
+};
 
-interface ColorBarLabelsProps {
-  minValue: number;
-  maxValue: number;
-  numTicks: number;
-  logScaled?: boolean;
-  onClick: (event: React.MouseEvent<HTMLDivElement>) => void;
-}
+const mapDispatchToProps = {};
 
-export default function ColorBarLabels({
-  minValue,
-  maxValue,
-  numTicks,
-  logScaled,
-  onClick,
-}: ColorBarLabelsProps) {
-  const labels = useMemo(
-    () => getLabelsForRange(minValue, maxValue, numTicks, logScaled),
-    [minValue, maxValue, numTicks, logScaled],
-  );
-  return (
-    <Box sx={styles.container} onClick={onClick}>
-      {labels.map((label, i) => (
-        <span key={i}>{label}</span>
-      ))}
-    </Box>
-  );
-}
+const MapPointInfoBox = connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(_MapPointInfoBox);
+export default MapPointInfoBox;